requirements.txt
python-telegram-bot
requests>=2.31.0
telegram
python-dotenv>=1.0.0
web3>=6.0
eth-account>=0.10.0
SQLAlchemy>=2.0
cryptography>=42.0
qrcode[pil]
Pillow>=10.0
feedparser>=6.0
APScheduler>=3.10
transformers>=4.41.0
torch>=2.3
prophet>=1.1
pandas>=2.2